Work In Progress: Create CategoryViewController and Category Entity
Current bug at time of this commit: Refactored code and created a Data Manager Singleton that handles all of the Core Data methods. Currently the app is functional; it doesn't crash. It is also successfully saving data to the persistent container, but I need to troubleshoot why the data isn't loading when app is relaunched.